[worker] Extract MessageQueueConsumer class in a separate file (#12623)
