+++ opencti-worker/src/worker.py
@@ -1,13 +1,9 @@
-# coding: utf-8
-
-import functools
 import os
 import signal
 import threading
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass, field
-from threading import Thread
-from typing import Any, Dict, List, Callable, Literal
+from typing import Any, Dict, List
 
 import pika
 import yaml
@@ -22,6 +18,7 @@ from pycti.connector.opencti_connector_helper import (
     get_config_variable,
 )
 
+from message_queue_consumer import MessageQueueConsumer
 from listen_handler import ListenHandler
 from push_handler import PushHandler
 
@@ -43,103 +40,6 @@ running_ingestion_units_gauge = meter.create_gauge(
     name="opencti_running_ingestion_units",
     description="Number of running ingestion units",
 )
-
-
-@dataclass(unsafe_hash=True)
-class MessageQueueConsumer:  # pylint: disable=too-many-instance-attributes
-    logger: Any
-    consumer_type: Literal["listen", "push"]
-    queue_name: str
-    pika_parameters: pika.ConnectionParameters
-    execution_pool: ThreadPoolExecutor
-    handle_message: Callable[[str], Literal["ack", "nack", "requeue"]]
-
-    def __post_init__(self) -> None:
-        self.should_stop = False
-        self.pika_connection = pika.BlockingConnection(self.pika_parameters)
-        self.channel = self.pika_connection.channel()
-        self.channel.basic_qos(prefetch_count=1)
-        self.thread = Thread(target=self.consume_queue, name=self.queue_name)
-        self.thread.start()
-
-    def nack_message(self, delivery_tag: int, requeue: bool) -> None:
-        if self.channel.is_open:
-            self.logger.info("Message rejected", {"tag": delivery_tag})
-            self.channel.basic_nack(delivery_tag, requeue=requeue)
-        else:
-            self.logger.info(
-                "Message NOT rejected (channel closed)", {"tag": delivery_tag}
-            )
-
-    def ack_message(self, delivery_tag: int) -> None:
-        if self.channel.is_open:
-            self.logger.info("Message acknowledged", {"tag": delivery_tag})
-            self.channel.basic_ack(delivery_tag)
-        else:
-            self.logger.info(
-                "Message NOT acknowledged (channel closed)",
-                {"tag": delivery_tag},
-            )
-
-    def consume_message(self, delivery_tag: int, body: str) -> None:
-        result = self.handle_message(body)
-        match result:
-            case "ack":
-                cb = functools.partial(self.ack_message, delivery_tag)
-                self.pika_connection.add_callback_threadsafe(cb)
-            case "nack":
-                cb = functools.partial(self.nack_message, delivery_tag, False)
-                self.pika_connection.add_callback_threadsafe(cb)
-            case "requeue":
-                cb = functools.partial(self.nack_message, delivery_tag, True)
-                self.pika_connection.add_callback_threadsafe(cb)
-
-    def consume_queue(self) -> None:
-        try:
-            self.logger.info(
-                "Thread for queue started",
-                {"consumer_type": self.consumer_type, "queue": self.queue_name},
-            )
-
-            # Consume the queue with a generator
-            for message in self.channel.consume(self.queue_name, inactivity_timeout=1):
-                if self.should_stop:
-                    break
-                if not all(message):
-                    continue
-                method, properties, body = message
-                self.logger.info(
-                    "Processing a new message, launching a thread...",
-                    {
-                        "consumer_type": self.consumer_type,
-                        "queue": self.queue_name,
-                        "tag": method.delivery_tag
-                    },
-                )
-                task_future = self.execution_pool.submit(
-                    self.consume_message,
-                    method.delivery_tag,
-                    body,
-                )
-                while task_future.running():  # Loop while the thread is processing
-                    self.pika_connection.sleep(0.05)
-                self.logger.info("Message processed, thread terminated")
-        except Exception as e:
-            self.logger.error("Unhandled exception", {"exception": e})
-        finally:
-            self.logger.info(
-                "Thread for queue terminated",
-                {"consumer_type": self.consumer_type, "queue": self.queue_name},
-            )
-
-    def is_alive(self) -> bool:
-        return self.thread.is_alive()
-
-    def stop(self, wait: bool) -> None:
-        self.should_stop = True
-        if wait:
-            self.thread.join()
-
 
 @dataclass(unsafe_hash=True)
 class Worker:  # pylint: disable=too-few-public-methods, too-many-instance-attributes
